pyproject.toml
[project]
name = "pdf-digitizer"
version = "0.1.0"
description = "Convert scanned PDFs to markdown using Marker OCR"
requires-python = ">=3.10"
dependencies = [
    "marker-pdf>=1.0.0",
]

[project.scripts]
digitize = "digitize:main"
